fix(gql): drop lossy float cast on base_state.dropout in gnn_train
TrainingState::dropout is declared as double in model_checkpoint.h
and serialized as 8 bytes. The static_cast<float> truncated the
user-configured dropout before writing it to the checkpoint, so the
persisted value silently differed from the training configuration.
Future tightening of check_arch_match to cover dropout would then
produce spurious mismatches on resume.

+++ src/query/procedure/builtin/gnn_train_procedure.cc
@@ -565,7 +565,7 @@ void GnnTrainProcedure::execute(ProcedureContext& ctx) {
     base_state.hidden_dim         = gnn_config.hidden_dim;
     base_state.num_classes        = gnn_config.num_classes;
     base_state.num_layers         = gnn_config.num_layers;
-    base_state.dropout            = static_cast<float>(gnn_config.dropout);
+    base_state.dropout            = gnn_config.dropout;
     base_state.normalize          = gnn_config.normalize;
     base_state.model_type         = model_type;   // "graphsage"
     base_state.projection_name    = projection_name;
